renames package package => packages
