+++ src/dimcat/data/datasets/base.py
@@ -25,8 +25,8 @@ from dimcat.data.base import Data
 from dimcat.data.catalogs.base import DimcatCatalog
 from dimcat.data.catalogs.inputs import InputsCatalog
 from dimcat.data.catalogs.outputs import OutputsCatalog
-from dimcat.data.package.base import Package, PackageSpecs
-from dimcat.data.package.dc import DimcatPackage
+from dimcat.data.packages.base import Package, PackageSpecs
+from dimcat.data.packages.dc import DimcatPackage
 from dimcat.data.resource.base import SomeDataframe
 from dimcat.data.resource.dc import DimcatResource
 from dimcat.data.resource.features import (
