+++ src/dimcat/data/packages/base.py
