+++ tests/conftest.py
@@ -11,7 +11,7 @@ import music21 as m21
 import pytest
 from _pytest.terminal import TerminalReporter
 from dimcat.data.datasets.base import Dataset
-from dimcat.data.package.dc import DimcatPackage
+from dimcat.data.packages.dc import DimcatPackage
 from dimcat.data.resource.dc import DimcatResource
 from dimcat.data.resource.utils import load_fl_resource, make_rel_path
 from dimcat.utils import scan_directory
