+++ tests/test_package.py
@@ -4,8 +4,8 @@ from typing import Type
 
 import pytest
 from dimcat.base import change_setting
-from dimcat.data.package.base import Package, PackageMode, PackageStatus, PathPackage
-from dimcat.data.package.dc import DimcatPackage
+from dimcat.data.packages.base import Package, PackageMode, PackageStatus, PathPackage
+from dimcat.data.packages.dc import DimcatPackage
 from dimcat.data.resource.base import PathResource
 from dimcat.dc_exceptions import ResourceIsFrozenError
